feat: finish mini cex

+++ lib/src/presentation/features/students/assesment/pages/mini_cex/student_mini_cex_detail.dart
@@ -56,7 +56,7 @@ class _StudentMiniCexDetailState extends State<StudentMiniCexDetail> {
                   TopStatCard(
                     title: 'Total Grades',
                     score: state.miniCexStudentDetail?.grade != null
-                        ? state.miniCexStudentDetail!.grade!.toInt()
+                        ? state.miniCexStudentDetail!.grade!.toDouble()
                         : 0,
                   ),
                   ...List.generate(
